+++ vkconfig_core/configuration.h
@@ -29,6 +29,8 @@
 #include <vector>
 #include <string>
 
+#define DEFAULT_PHYSICAL_DEVICE "Default Physical Device"
+
 class LayerManager;
 
 class Configuration {
@@ -63,9 +65,11 @@ class Configuration {
     bool override_settings = false;
     Path override_settings_path;
     bool override_layers = true;
+    std::string selected_layer_name;
+    bool override_driver = false;
+    std::string override_driver_name = DEFAULT_PHYSICAL_DEVICE;
     bool override_loader = true;
     int loader_log_messages_flags = GetBit(LOG_ERROR);
-    std::string selected_layer_name;
 
     std::vector<Parameter> parameters;