+++ vkconfig_gui/tab_preferences.cpp
@@ -158,6 +158,9 @@ void TabPreferences::on_theme_mode_changed(int index) {
             ? "Disable external layers settings file, use Vulkan Configurator generated layers settings file"
             : "Hide 'Use Layers Settings file' bar by default");
 
+    // Drivers
+    this->ui->driver_browse_button->setIcon(::Get(new_theme_mode, ::ICON_FOLDER_SEARCH));
+
     // Layers
     this->ui->layers_browse_button->setIcon(::Get(new_theme_mode, ::ICON_FOLDER_SEARCH));
     this->ui->layers_reload_button->setIcon(::Get(new_theme_mode, ::ICON_FOLDER_RELOAD));